# Remove unnecessary allocation in `BitSet()`