@LilithHafner 2023-04-22
This works with `_growbeg!`, too, which is relevant for `BitSet`

```julia
julia> @btime begin x = sizehint!(zeros(UInt64, 0), 4); for _ in 1:3; Base._growbeg!(x, 1); end; end
  52.379 ns (2 allocations: 144 bytes)

julia> @btime begin x = resize!(Vector{UInt64}(undef, 4), 0); for _ in 1:3; Base._growbeg!(x, 1); end; end
  43.264 ns (1 allocation: 96 bytes)
```