@vtjnash 2023-04-22
The benchmarks are unfair there, since you never use a large enough workload to show the higher wasted memory of this change. It is not large waste, so it might be fine if BitSet is normally less than 512. Otherwise this may be worse.