@vtjnash 2023-04-22
resized memory can be freed later. The initial allocation cannot be freed later